grpc-python/requirements.txt
autopep8==1.5.6
binaryornot==0.4.4
boolean.py==3.8
certifi==2020.12.5
chardet==4.0.0
grpcio==1.37.0
grpcio-tools==1.37.0
idna==2.10
Jinja2==2.11.2
license-expression==1.2
MarkupSafe==1.1.1
protobuf==3.15.8
pycodestyle==2.7.0
python-debian==0.1.39
requests==2.25.1
reuse==0.12.1
six==1.15.0
toml==0.10.2
urllib3==1.26.2
